+++ src/public-api.ts
@@ -18,6 +18,7 @@ export * from './lib/errors/invalid-limit.error';
 export * from './lib/errors/invalid-page-number.error';
 export * from './lib/errors/invalid-resource-name.error';
 export * from './lib/errors/key-not-found.error';
+export * from './lib/errors/pagination-not-synced.error';
 export * from './lib/errors/unselectable-model.error';
 export * from './lib/errors/unsupported-field-selection.error';
 export * from './lib/errors/unsupported-filter.error';
@@ -41,6 +42,9 @@ export * from './lib/interfaces/query-builder-state.interface';
 export * from './lib/interfaces/request-strategy.interface';
 export * from './lib/interfaces/response-strategy.interface';
 export * from './lib/interfaces/sort.interface';
+
+// Injection tokens
+export * from './lib/tokens/ng-qubee.tokens';
 
 // Strategies
 export * from './lib/strategies/json-api-request.strategy';